refactor: replace FixedInfiniteVirtualList with VariableInfiniteVirtualList in tests and components
This update transitions the VirtualList component to utilize VariableInfiniteVirtualList, enhancing its functionality for handling variable row heights. The changes include updates to test cases and component imports to reflect this new implementation, ensuring improved performance and user experience in infinite scrolling scenarios.
